refactor: move prop generators to gossip module
Move the arbitrary namespace / table / column generators used for
property testing to the parent "gossip" module, so they can be shared
across the sync implementation tests too.
